pyproject.toml
[project]
name = "gf-2o"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "elevenlabs>=2.26.1",
    "groq>=0.37.1",
    "langchain>=1.1.3",
    "langchain-groq>=1.1.1",
    "pydantic-settings>=2.12.0",
    "together>=1.5.32",
]
